Add few key tests and modify structure
